+++ util/retry_test.go
@@ -1,8 +1,11 @@
 package util_test
 
 import (
+	"fmt"
 	"testing"
 	"time"
+
+	"github.com/pkg/errors"
 
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/cycloidio/terracognita/util"
@@ -50,5 +53,28 @@ func TestRetry(t *testing.T) {
 		err := util.Retry(fn, 3, 0*time.Second)
 		require.Equal(t, err, awserr.New(throttlingErr, "message", nil))
 		assert.Equal(t, 3, count)
+	})
+	t.Run("NoRetrySTD", func(t *testing.T) {
+		var count int
+		fn := func() error {
+			count++
+			return fmt.Errorf("some std error")
+		}
+
+		err := util.Retry(fn, 3, 0*time.Second)
+		require.Equal(t, err, fmt.Errorf("some std error"))
+		assert.Equal(t, 1, count)
+	})
+	t.Run("NoRetry|pk/errors", func(t *testing.T) {
+		var count int
+		eerr := errors.New("some custom error")
+		fn := func() error {
+			count++
+			return eerr
+		}
+
+		err := util.Retry(fn, 3, 0*time.Second)
+		require.Equal(t, eerr, err)
+		assert.Equal(t, 1, count)
 	})
 }